chore(urls): Fix test failures (#3382)
