+++ hawtio-system/src/main/java/io/hawt/web/ServletHelpers.java
@@ -7,6 +7,8 @@ import java.io.PrintWriter;
 import java.net.URL;
 import javax.management.AttributeNotFoundException;
 
+import io.hawt.web.auth.SessionExpiryFilter;
+import jakarta.servlet.ServletContext;
 import jakarta.servlet.http.HttpServletResponse;
 
 import io.hawt.system.Authenticator;
@@ -144,4 +146,90 @@ public class ServletHelpers {
         }
         return header.replaceAll("[\\r\\n]", "");
     }
+
+    /**
+     * Normalizes a path. If the path contains a single '/' character it is returned
+     * unchanged, otherwise the path is:
+     * <ol>
+     * <li>stripped from all multiple consecutive occurrences of '/' characters</li>
+     * <li>stripped from trailing '/' character(s)</li>
+     * </ol>
+     *
+     * @param path
+     *            path to normalize
+     * @return normalized path
+     */
+    public static String cleanPath(final String path) {
+        final String result = path.replaceAll("//+", "/");
+        return result.length() == 1 && result.charAt(0) == '/' ? result
+                : result.replaceAll("/+$", "");
+    }
+
+    /**
+     * Creates a web context path from components. Concatenates all path components
+     * using '/' character as delimiter and the result is then:
+     * <ol>
+     * <li>prefixed with '/' character</li>
+     * <li>stripped from all multiple consecutive occurrences of '/' characters</li>
+     * <li>stripped from trailing '/' character(s)</li>
+     * </ol>
+     *
+     * @return empty string or string which starts with a "/" character but does not
+     *         end with a "/" character
+     */
+    public static String webContextPath(final String first, final String... more) {
+        if (more.length == 0 && (first == null || first.isEmpty())) {
+            return "";
+        }
+
+        final StringBuilder b = new StringBuilder();
+        if (first != null) {
+            if (!first.startsWith("/")) {
+                b.append('/');
+            }
+            b.append(first);
+        }
+
+        for (final String s : more) {
+            if (s != null && !s.isEmpty()) {
+                b.append('/');
+                b.append(s);
+            }
+        }
+
+        final String cleanedPath = cleanPath(b.toString());
+        return cleanedPath.length() == 1 ? "" : cleanedPath;
+    }
+
+    /**
+     * Return a number of web path segments that need to be skipped to reach <em>hawtio path</em>. In JakartaEE
+     * environment (WAR) everything after context path is "hawtio path", so {@code 0} is returned. In Spring Boot
+     * we may have to skip some segments (like {@code /actuator/hawtio}).
+     *
+     * @param servletContext
+     * @return
+     */
+    public static int hawtioPathIndex(ServletContext servletContext) {
+        String servletPath = (String) servletContext.getAttribute(SessionExpiryFilter.SERVLET_PATH);
+        if (servletPath == null) {
+            // this attribute is set only in non JakartaEE environments, so here we are in standard WAR
+            // deployment. Just return "0", which means full path without initial context path
+            return 0;
+        } else {
+            // when SessionExpiryFilter.SERVLET_PATH is set, it contains prefix which should be skipped and which
+            // is not standard JakartaEE path components (context path, servlet path, path info).
+            // for Spring Boot we have to skip dispatcher servlet path, management endpoints base ("/actuator")
+            // and management endpoint mapping
+            String cleanPath = webContextPath(servletPath);
+            int pathIndex = 0;
+            // for "/actuator/hawtio", we have to return "2", so count slashes
+            for (char c : cleanPath.toCharArray()) {
+                if (c == '/') {
+                    pathIndex++;
+                }
+            }
+            return pathIndex;
+        }
+    }
+
 }
